+++ pom.xml
@@ -31,7 +31,7 @@
             <plugin>
                 <groupId>org.jetbrains.kotlin</groupId>
                 <artifactId>kotlin-maven-plugin</artifactId>
-                <version>1.5.10</version>
+                <version>1.5.31</version>
                 <executions>
                     <execution>
                         <id>compile</id>
@@ -52,7 +52,7 @@
             <plugin>
                 <groupId>org.codehaus.mojo</groupId>
                 <artifactId>exec-maven-plugin</artifactId>
-                <version>1.6.0</version>
+                <version>3.0.0</version>
                 <configuration>
                     <mainClass>MainKt</mainClass>
                 </configuration>
@@ -63,8 +63,8 @@
     <dependencies>
         <dependency>
             <groupId>org.jetbrains.kotlin</groupId>
-            <artifactId>kotlin-stdlib-jdk8</artifactId>
-            <version>1.5.10</version>
+            <artifactId>kotlin-stdlib</artifactId>
+            <version>1.5.31</version>
         </dependency>
         <dependency>
             <groupId>javazoom</groupId>
@@ -74,7 +74,7 @@
         <dependency>
             <groupId>org.jetbrains.kotlinx</groupId>
             <artifactId>kotlinx-coroutines-core</artifactId>
-            <version>1.3.8</version>
+            <version>1.5.2</version>
         </dependency>
 
     </dependencies>
