--- src/main/kotlin/main.kt
@@ -1,137 +1,0 @@
-import javazoom.jl.player.Player
-import kotlinx.coroutines.*
-import java.io.InputStream
-import java.net.URI
-import java.net.URLEncoder
-import java.net.http.HttpClient
-import java.net.http.HttpRequest
-import java.net.http.HttpResponse
-
-const val generateSoundUrl = "https://nextup.com/ivona/php/nextup-polly/CreateSpeech/CreateSpeechGet3.php?voice=Maxim&language=ru-RU&text="
-
-val delays = longArrayOf(
-    400,
-    400,
-    800,
-    450,
-    200,
-    700,
-    500,
-    1000,
-    600,
-    600,
-    700,
-    500,
-    900,
-    600,
-    600,
-    600,
-    700,
-    500,
-    400,
-    400,
-    600,
-    600
-)
-val textLines = listOf(
-    "А, мой, мальчик, едет на девятке\n",
-    "По автостраде, вдоль ночных дорог\n",
-    "Я, круужилась, с ним на танцплощадке\n",
-    "А ты и дальше, будешь одинок\n"
-)
-
-fun main() = runBlocking {
-    val words = textLines.flatMap { it.split(" ") }
-    val sounds = with(HttpClient.newBuilder().build()) {
-        textLines.map { line ->
-            val generateSoundRequest = HttpRequest.newBuilder()
-                .uri(URI.create(generateSoundUrl + URLEncoder.encode(line, "UTF-8")))
-                .build()
-            val soundUrl = send(
-                generateSoundRequest,
-                HttpResponse.BodyHandlers.ofString()
-            ).body()
-            val playSoundRequest = HttpRequest.newBuilder()
-                .uri(URI.create(soundUrl))
-                .build()
-
-            send(
-                playSoundRequest,
-                HttpResponse.BodyHandlers.ofInputStream()
-            ).body()
-        }
-    }
-
-    playAndWrite(sounds, words)
-    printPhoto()
-}
-
-suspend fun playAndWrite(
-    sounds: List<InputStream>,
-    words: List<String>
-) = withContext(Dispatchers.IO) {
-    listOf(
-        launch {
-            write(words)
-        },
-        launch {
-            sounds.forEach { sound ->
-                Player(sound).play()
-                delay(400L)
-            }
-        }
-    ).joinAll()
-}
-
-suspend fun write(words: List<String>) {
-    words.onEach { word -> print("$word ") }
-        .mapIndexedNotNull { index, _ -> delays.getOrNull(index) }
-        .forEach { currentDelay -> delay(currentDelay) }
-}
-
-suspend fun printPhoto() {
-    arrayOf(
-        "+++++++++++++==+++++++++++++++++++++================+======+=+===+============++++++=++++++++++++++++====++=============",
-        "+++++++++++++++++++=+++++++++==========================%%%%%%===========================================================",
-        "+++++++++++++++++++++=++++++====================%@@###############@%====================================================",
-        "++++++++++++++===+======+++==================@################@#######%=================================================",
-        "++++++++++==+===+++++=+++=+===============%@############################@%==============================================",
-        "+++++++++=======++++=+++++==============%@################################@%============================================",
-        "+++++=+======++++====++==+=============@########################@@##########%%==========================================",
-        "++++=++=+====+======+++==============%####################%=======%@@@#######@%=========================================",
-        "+++++++==++=++===+++====+===========%###################%=+++++++++==%@########%========================================",
-        "+++++++==+++=======================%##################@%+******++++++=%@########%=======================================",
-        "++++++++++++=+====================%#################@%=+*********++++++%#########@%=====================================",
-        "++++++++++++=====================%@###############@%==+***********+++++=@##########%====================================",
-        "++++++++++=======================@###############@==++*************++++=%@##########%===================================",
-        "+++++++++=======================@###############@==+*******:******+==%%==%###########@%=================================",
-        "++++++++++++======+============@@##############@@%=++********+%@@%========%###########@%================================",
-        "+++++++++++==================%%%###############@%=====++****+==%@#####%++=%############@%===============================",
-        "+++++++++++=================%%%%##############@%####@%=++**+====+=%%%=+++=%@###########@%===============================",
-        "+++++++++++===============%%%%%###############%%@%=%%==+++*+==++++*****++=%@############@%==============================",
-        "++++++++++===============%@@@#################@===+++++*++*+==++*******++=%@############@%==============================",
-        "+++++++++==============%%%####################@=+++******+++=%=+******++=%@##############@%=============================",
-        "+++++++++=============%%@######################%++******++++==%+******+==%@##############@%=============================",
-        "+++++++++===========%%@#########################=+*******+++=%=+*****++=%%@###############@%============================",
-        "+++++++++++========%=%###########################=+*******+**+++*****++=%%@################@@%==========================",
-        "+++++++++++========%=@###########################@=+******++===+++**+++==%##################@@@%========================",
-        "++++++++++++=======%%@#############################=+***+=%%==%@%%++++==%#####################@@@%======================",
-        "+++++++++++=========%###############################%+***+========+++==%#######################@@@@@@@%=================",
-        "+++++++++=========%###################################=+*****++*++++=%%@#########################@@@@@@@%===============",
-        "+++++++++=======@#######################################=********+====%%@#########################@@@@@@@@%=============",
-        "++++++++========%#######################################@%=+++++=======%@##########################@@@@@@@%=============",
-        "+++++++========%%######################################@@%==============@###########################@@@@@@%=============",
-        "++++++========%@#####################################@@@%%==============@#############################@@@@@%============",
-        "+++++++======%@#####################################@%%%%==============%###############################@@@@@@@@%========",
-        "+++++++=====%@######################################%%=%%====+++++++===%################################@@@@@@@@%=======",
-        "++++++===%%#########################################%======++++++++++++%################################@@@@@@@%@%======",
-        "++++++==%#@#########################################%====%==++++++++++=@#################################@@@@@@%@%======",
-        "+++++===@##########################################@%=================+##################################@@@@@@@@%======",
-        "+++++==%###########################################@====++++++++++++==%##################################@@@@@@@@@@@%===",
-        "+++++=%############################################@==++++++++++++++++####################################@@@@@@@@@@%===",
-        "++++==#############################################@=++++++*****+++++%#####################################@@@@@@@@@%==="
-    ).forEach {
-        println(it)
-        delay(75)
-    }
-}
